Underpin custom properties, initial commit
The model is still kind of hardcoded, but easier to extend than before.
The custom properties are written on the customProperties field on the connector so no additional changes needed on the connector or the database schema.

To be further extended, we keep this now for the demo and rework afterwards.